Visualize the integration of bulk DE results with scRNA-seq datasets

This function assists with visualization of the integration of bulk DE results tables with pre-processed scRNA-seq datasets available on Bioconductor. After a user has picked a scRNA-seq dataset using integrateWithSingleCell from the DESeq2 package, they will have a results table, DESeqDataSet, and a SingleCellExperiment. After annotating the SingleCellExperiment (see vignette), they can provide that list of object to this function. This function then produces a violin plot of the expression of genes of interest from the bulk DE analysis across annotated cell types in the single-cell dataset.

vizWithSCE(dat, which)

Arguments

dat
the output of integrateWithSingleCell, a list with results table (res), DESeqDataSet (dds), and SingleCellExperiment (sce)

which
which gene to use for the violin plot. This is represented as a number that corresponds to the lowest p-value e.g. 1 for the lowest p-value, 2 for the second lowest p-value, etc.

Value

a violin plot (see Description)

Author

Kwame Forbes

Michael Love

Examples

if (FALSE) {
dat <- integrateWithSingleCell(res,dds)
vizWithSCE(dat, which=1)
}